refactor: 서비스 계층에 있는 애플 탈퇴 코드를 AppleLoginStrategy로 옮김 & 매개변수 일부 수정
