--- BE/src/auth/dto/delete-auth.dto.ts
@@ -1,12 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsString } from 'class-validator';
-
-export class DeleteAuthDto {
-  @ApiProperty()
-  @IsString()
-  idToken: string;
-
-  @ApiProperty()
-  @IsString()
-  authorizationCode: string;
-}
